Masterscript/requirements.txt
wandb
tensorboard
albumentations>=1.1.0
p_tqdm==1.3.3
pytest
pytest-cov
tensorflow
imageio==2.16.2
keras
opencv_python_headless>=4.5.5.62
elephant==0.12.0
joblib==1.2.0
numpy>=1.21.2
opencv-python==4.7.0.72
Pillow==9.5.0
PyYAML>=5.4.1
scikit-learn==1.2.2
scipy>=1.10.1
torch>=1.10.0
torchelastic==0.2.2
torchtext
torchvision>=0.11.0
webcolors==1.13
yacs==0.1.8
zipp==3.15.0
openpyxl
efficientnet_pytorch>=0.7.1
matplotlib>=3.2.2
prefetch_generator>=1.0.1
pretrainedmodels>=0.7.4
tensorboardX>=2.4.1
timm>=0.5.4
tqdm>=4.61.2
webcolors>=1.11.1
pandas
plotly
kaleido